COLORED LIGHTING!!!! Huge performance improvements aswell

+++ src/com/dazkins/triad/game/entity/LightEmitter.java
@@ -2,5 +2,5 @@ package com.dazkins.triad.game.entity;
 
 //TODO: Make this class abstract not interface
 public interface LightEmitter {
-	public byte getLightStrength();
+	public float getLightStrength();
 }
